Allow both values of type String and of type Id to appear in $id: antiquotes.
Values that appear in a $id: antiquote must be of a type for which there is an
instance of the ToIdent class. Language.C.Quote.Base defines instanced for
String, Id, and SrcLoc -> Id. The ToIdent class is analogous to the ToExp class
which is used to specify which types may appear in a $exp: antiquotation.