Fix pretty-print of dangling else statements.
We previously didn't always pretty-print dangling else's in nested if
statements such that they associated with the proper if. Now we do.

To ensure that prop_ParsePrintUnitId holds, we now also normalize parsed
statement blocks so that if they contain a single statement, they parse
as a single statement rather than a statement block consisting of a
single statement. This required some other fixes in the pretty printer.

In practice, we never hit this problem in parsed code, but we could
potentially encounter it in generated code.